movie の id 生成を action に移動
